+++ qcluster/__init__.py
@@ -1,4 +1,4 @@
 from pathlib import Path
 
 ROOT_DIR = Path(__file__).parent
-__version__ = "0.1.5"
+__version__ = "0.1.6"