+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [0.1]
 
+### [0.1.6]
+- Add cluster describer, using LLMs to generate descriptions and titles
+- Integrate with `InstructionCollection`
+
 ### [0.1.5]
 - Add dissimilarity module and integrate with `InstructionCollection`